# public/.well-known/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "YAZIO Meal Logger",
    "description": "Log meals, search foods, and view nutrition summaries via YAZIO.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://your-app.vercel.app"
    }
  ],
  "paths": {
...
    "/api/tools/search-food": {
      "get": {
        "operationId": "searchFood",
        "summary": "Search for foods in YAZIO database",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": { "type": "string" },
            "description": "Food name to search for (German)"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 5, "maximum": 20 },
            "description": "Max results"
          }
        ],
        "responses": {
          "200": {
            "description": "Search results",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "results": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "string" },
                          "name": { "type": "string" },
                          "brand": { "type": "string" },
                          "calories": { "type": "number" },
                          "protein": { "type": "number" },
                          "carbs": { "type": "number" },
                          "fat": { "type": "number" }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ]
}